--- data/src/main/java/com/devndev/data/datsource/ItemDataSource.kt
@@ -1,7 +1,0 @@
-package com.devndev.data.datsource
-
-import com.devndev.domain.model.Item
-
-interface ItemDataSource {
-    suspend fun fetchItems(): List<Item>
-}
