+++ data/src/main/java/com/devndev/lamp/data/datsource/ItemDataSource.kt
@@ -1,0 +1,7 @@
+package com.devndev.lamp.data.datsource
+
+import com.devndev.lamp.domain.model.Item
+
+interface ItemDataSource {
+    suspend fun fetchItems(): List<Item>
+}
