+++ data/src/main/java/com/devndev/lamp/data/di/RepositoryModule.kt
@@ -1,0 +1,17 @@
+package com.devndev.lamp.data.di
+
+import com.devndev.lamp.data.repository.ItemRepositoryImpl
+import com.devndev.lamp.domain.repository.ItemRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+import javax.inject.Singleton
+
+@Module
+@InstallIn(SingletonComponent::class)
+interface RepositoryModule {
+    @Singleton
+    @Binds
+    fun bindItemRepository(itemRepositoryImpl: ItemRepositoryImpl): ItemRepository
+}
