--- domain/src/main/java/com/devndev/domain/model/Item.kt
@@ -1,6 +1,0 @@
-package com.devndev.domain.model
-
-data class Item(
-    val id: Int,
-    val name: String
-)
