--- domain/src/main/java/com/devndev/domain/repository/ItemRepository.kt
@@ -1,7 +1,0 @@
-package com.devndev.domain.repository
-
-import com.devndev.domain.model.Item
-
-interface ItemRepository {
-    suspend fun getItems(): List<Item>
-}
