+++ domain/src/main/java/com/devndev/lamp/domain/model/Item.kt
@@ -1,0 +1,6 @@
+package com.devndev.lamp.domain.model
+
+data class Item(
+    val id: Int,
+    val name: String
+)
