+++ domain/src/main/java/com/devndev/lamp/domain/repository/ItemRepository.kt
@@ -1,0 +1,7 @@
+package com.devndev.lamp.domain.repository
+
+import com.devndev.lamp.domain.model.Item
+
+interface ItemRepository {
+    suspend fun getItems(): List<Item>
+}
