+++ presentation/src/main/java/com/devndev/lamp/presentation/ui/s2/Screen2.kt
@@ -1,0 +1,9 @@
+package com.devndev.lamp.presentation.ui.s2
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun Screen2() {
+    Text("Screen2")
+}
