+++ presentation/src/main/java/com/devndev/lamp/presentation/ui/s3/Screen3.kt
@@ -1,0 +1,9 @@
+package com.devndev.lamp.presentation.ui.s3
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun Screen3() {
+    Text("Screen3")
+}
