+++ presentation/src/main/java/com/devndev/lamp/presentation/ui/s4/Screen4.kt
@@ -1,0 +1,9 @@
+package com.devndev.lamp.presentation.ui.s4
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+fun Screen4() {
+    Text("Screen4")
+}
