--- presentation/src/main/java/com/devndev/presentation/main/MainActivity.kt
@@ -1,45 +1,0 @@
-package com.devndev.presentation.main
-
-import android.os.Bundle
-import android.util.Log
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.setContent
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.ui.Modifier
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.compose.rememberNavController
-import com.devndev.presentation.ui.login.LoginScreen
-import com.devndev.presentation.ui.login.LoginViewModel
-import com.devndev.presentation.ui.splsh.SplashScreen
-import com.devndev.presentation.ui.theme.LampTheme
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : ComponentActivity() {
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContent {
-            LampTheme {
-                Lamp()
-            }
-        }
-    }
-}
-
-@Composable
-fun Lamp(viewModel: LoginViewModel = hiltViewModel()) {
-    val navController = rememberNavController()
-    val isLoggedIn = viewModel.isLoggedIn.collectAsState().value
-    val isInitialized = viewModel.isInitialized.collectAsState().value
-    Log.d("----", "isLoggedIn $isLoggedIn")
-    if (isInitialized) {
-        if (isLoggedIn) {
-            MainScreen(modifier = Modifier)
-        } else {
-            LoginScreen(modifier = Modifier)
-        }
-    } else {
-        SplashScreen()
-    }
-}
