--- presentation/src/main/java/com/devndev/presentation/ui/home/HomeScreen.kt
@@ -1,37 +1,0 @@
-package com.devndev.presentation.ui.home
-
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.unit.dp
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.devndev.domain.model.Item
-
-@Composable
-fun HomeScreen(viewModel: HomeViewModel = hiltViewModel(), modifier: Modifier) {
-    val items: List<Item> by viewModel.items.collectAsState()
-
-    Column(
-        modifier = modifier
-            .fillMaxSize()
-            .padding(16.dp),
-        verticalArrangement = Arrangement.Center,
-        horizontalAlignment = Alignment.CenterHorizontally
-    ) {
-        Text(text = "Home Screen", style = MaterialTheme.typography.bodyMedium)
-        Spacer(modifier = Modifier.height(16.dp))
-        items.forEach { item ->
-            Text(text = item.name, style = MaterialTheme.typography.bodyMedium)
-        }
-    }
-}
