--- presentation/src/main/java/com/devndev/presentation/ui/home/HomeViewModel.kt
@@ -1,30 +1,0 @@
-package com.devndev.presentation.ui.home
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.devndev.domain.model.Item
-import com.devndev.domain.usecase.ItemUseCase
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.launch
-import javax.inject.Inject
-
-@HiltViewModel
-class HomeViewModel @Inject constructor(
-    private val itemUseCase: ItemUseCase
-) : ViewModel() {
-
-    private val _items = MutableStateFlow<List<Item>>(emptyList())
-    val items: StateFlow<List<Item>> = _items
-
-    init {
-        fetchData()
-    }
-
-    private fun fetchData() {
-        viewModelScope.launch {
-            _items.value = itemUseCase()
-        }
-    }
-}
