--- presentation/src/main/java/com/devndev/presentation/ui/s2/Screen2.kt
@@ -1,9 +1,0 @@
-package com.devndev.presentation.ui.s2
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun Screen2() {
-    Text("Screen2")
-}
