--- presentation/src/main/java/com/devndev/presentation/ui/s3/Screen3.kt
@@ -1,9 +1,0 @@
-package com.devndev.presentation.ui.s3
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun Screen3() {
-    Text("Screen3")
-}
