--- presentation/src/main/java/com/devndev/presentation/ui/s4/Screen4.kt
@@ -1,9 +1,0 @@
-package com.devndev.presentation.ui.s4
-
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-
-@Composable
-fun Screen4() {
-    Text("Screen4")
-}
