--- presentation/src/main/java/com/devndev/presentation/ui/splsh/SplashScreen.kt
@@ -1,18 +1,0 @@
-package com.devndev.presentation.ui.splsh
-
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-
-@Composable
-fun SplashScreen() {
-    Box(
-        modifier = Modifier.fillMaxSize(),
-        contentAlignment = Alignment.Center
-    ) {
-        Text("Splash Screen")
-    }
-}
